backend/requirements.txt
Flask
pandas
openpyxl
fpdf2
matplotlib
